requirements.txt
# Backend server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiofiles>=23.2.0
python-multipart>=0.0.9

# Audio processing
librosa>=0.10.0
numpy>=1.26.0
scipy>=1.12.0
pydub>=0.25.1
soundfile>=0.12.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# DJ Commentary Engine
ollama>=0.4.0

# Data
pandas>=2.2.0
requests>=2.31.0
tqdm>=4.66.0
yt-dlp>=2023.12.30
replicate>=1.0.7
